e2fsck: fix potential Floating Point Exception in show_stats()
If the free inodes count in the superblock is equal to the inodes
count in the superblock (which is not possible with a valid file
system and will be fixed by e2fsck unless it is prevented by, for
example, e2fsck -n), it is possible for e2fsck to crash due to a
divide by zero error.

Fix this potential bug.

Addresses-Debian-Bug: #879220

Signed-off-by: Theodore Ts'o <tytso@mit.edu>