configure: don't enable LTO by default
Not all compilers (or versions of compilers) use the same LTO options
or behave the same way with LTO.  In particular, using clang and the
current LTO options cause the build to fail.  We should probably fix
up the configure script to handle Clang and LTO, but for now, we won't
enable LTO unless the user explicitly passes --enable-lto to the
configure script.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>