Fuse2fs: add fakeroot option.
Add a new 'fakeroot' option to fuse2fs. When enabled, fuse2fs will
will pretend to be root when checking file permssions. This allows
fuse2fs to be used for building/modifying rootfs images as an
unprivileged user.

Nosuid and nodev are enabled automatically when fakeroot is enabled
as per maintainer's request, assuming they're supported on the target
platform, detected at ./configure time (at the time of this writing,
FreeBSD and possibly others don't provide a nodev mount option).

Signed-off-by: Nicholas Clark <nicholas.clark@gmail.com>