ext2fs: nls: support UTF-8 11.0 with NFKD normalization
We need this such that we can do normalization and casefolding
compatible with the kernel, in order to properly support fsck
verification and rehashing.

The UTF-8 11.0 implementation is copied and adapted from the kernel code
to ensure maximum compatibility.  The decode trie in utf8data.h is
generated using a script and the UCD sources in the kernel code.

Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>