Convert kernel compat functions to use new ll_rw_block() function signature
In newer kernels, ll_rw_block() separated the request operation and
the operational flags arguments.  This means adding a new parameter to
ll_rw_block() (which is ignored in our compat layer) and changing READ
and WRITE to REQ_OP_READ and REQ_OP_WRITE, respectively.

This makes it easier to keep us in sync with the kernel tree.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>