Rename functions, types, constants to reflect jbd2 usage
We had previously stuck to using the names from ext3/jbd kernel files,
and used a script in contrib/jbd2-resync.sh to convert the kernel
files to use the ext3/jbd conventions so we could keep the files
e2fsck/recovery.c and e2fsck/revoke.c in sync with jbd2/recovery.c and
jbd2/revoke.c, respectively.

This has been getting harder and harder, so let's make a global sweep
through e2fsprogs to use the jbd2 names.  Fortunately none of the
ext3/jbd names had leaked out into publically exported header files,
so this is only an internal change.  Which looks scary, but it's
basically a search and replace, so if it compiles it's going to be
correct.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>