libext2fs: use __u64 instead of ext2_off64_t in qcow.c
The type ext2_off64_t should only be used in the context of the ext2fs
library's llseek() and ftruncate() analogs, ext2_file_llseek() and
ext2_inode_set_size().  That's because like the POSIX.1's off_t, it
needs to be a signed integer value so that SEEK_END will work
correctly.

qcow.c's use of ext2_off64_t is only internal, and not in any exposed
interfaces, so changes nothing; it just prepares us to be able to
change the type of ext2_off64_t.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>