tst_bitops: don't stop testing on low memory
Some build systems have 512 MiB memory or less, e.g. Ubiquity Edgeroute
Lite which can be used to build for the mips and mips64 architectures.

On these allocating 512 MIB will always fail. So if allocating 512 MiB
fails treat this as expected behavior and return 0.

This will allow testing to continue with the remaining tests.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>